переместил файл

+++ scrapper/src/main/java/ru/tinkoff/scrapper/dto/responce/link/LinkResponse.java
@@ -2,5 +2,5 @@ package ru.tinkoff.scrapper.dto.responce.link;
 
 import java.net.URI;
 
-public record LinkResponse(Integer id, URI url) {
+public record LinkResponse(Long id, URI url) {
 }
